apps: rename build-deploy-container
