+++ aws/06-apps/build-deploy-container/index.ts
@@ -12,39 +12,25 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import * as azure from "@pulumi/azure";
-import * as docker from "@pulumi/docker";
+import * as awsx from "@pulumi/awsx";
 import * as k8s from "@pulumi/kubernetes";
-import * as pulumi from "@pulumi/pulumi";
 import { config } from "./config";
 
-// Create an Azure Resource Group
-const resourceGroup = new azure.core.ResourceGroup("samples");
+// Create a repository.
+const repo = new awsx.ecr.Repository("my-repo");
 
-const registry = new azure.containerservice.Registry("myregistry", {
-    resourceGroupName: resourceGroup.name,
-    sku: "Basic",
-    adminEnabled: true,
-});
-
+// Build a Docker image from a local Dockerfile context in the
+// './node-app' directory, and push it to the registry.
 const customImage = "node-app";
-const appImage = new docker.Image(customImage, {
-    imageName: pulumi.interpolate`${registry.loginServer}/${customImage}:v1.0.0`,
-    build: {
-        context: `./${customImage}`,
-    },
-    registry: {
-        server: registry.loginServer,
-        username: registry.adminUsername,
-        password: registry.adminPassword,
-    },
-});
+const appImage = repo.buildAndPushImage(`./${customImage}`);
 
+// Create a k8s provider.
 const provider = new k8s.Provider("provider", {
     kubeconfig: config.kubeconfig,
     namespace: config.appsNamespaceName,
 });
 
+// Create a Deployment of the built container.
 const appLabels = { app: customImage };
 const appDeployment = new k8s.apps.v1.Deployment("app", {
     spec: {
@@ -55,7 +41,7 @@ const appDeployment = new k8s.apps.v1.Deployment("app", {
             spec: {
                 containers: [{
                     name: customImage,
-                    image: appImage.imageName,
+                    image: appImage,
                     ports: [{name: "http", containerPort: 80}],
                 }],
             }
