+++ aws/06-apps/build-deploy-container/tsconfig.json
