+++ azure/06-apps/build-deploy-container/Pulumi.yaml
