+++ azure/06-apps/build-deploy-container/config.ts
