+++ azure/06-apps/build-deploy-container/node-app/app/package.json
