+++ azure/06-apps/build-deploy-container/package.json
