+++ gcp/06-apps/build-deploy-container/node-app/.dockerignore
